requirements.txt
python-dotenv>=1.0.0
langgraph>=0.2.0
pyyaml>=6.0.0
jsonschema>=4.0.0
openai>=1.0.0